animepahe2/backend/requirements.txt
fastapi>=0.100.0,<0.101.0
uvicorn[standard]>=0.22.0,<0.23.0
pydantic>=2.0.0,<3.0.0
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
selenium>=4.10.0,<5.0.0
undetected-chromedriver>=3.5.0,<4.0.0
Pillow>=10.0.0,<11.0.0
tenacity>=8.2.0,<9.0.0
jsbeautifier>=1.14.0,<2.0.0